// src/app/@api/model/invoiceRequestDto.ts
/**
 * apis backend
 * Ahmed Abdelrhman & Mohamed sabri
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { InquiryVersionIdDto } from './inquiryVersionIdDto';
import { OrderIdDto } from './orderIdDto';
import { QuotationVersionIdDto } from './quotationVersionIdDto';


export interface InvoiceRequestDto { 
    comment?: string;
    dealId: number;
    inquiryVersion?: InquiryVersionIdDto;
    order?: OrderIdDto;
    quotationVersion?: QuotationVersionIdDto;
}